No ramo master  Alterações para serem submetidas: 	modificado:    "Declaração.h" 	novo ficheiro: Makefile 	modificado:    aberturadeconta.c 	modificado:    auxiliar.c 	modificado:    cliente.c 	modificado:    data.c 	modificado:    essentials.c 	modificado:    "histórico.c" 	modificado:    kid.c 	modificado:    lscli.c 	modificado:    lshst.c 	modificado:    lskid.c 	modificado:    main.c 	modificado:    menus.c
Versão 1.0:	Modularização do Projeto.
		Criação do Makefile.
		Funcionamento de todas as funções (Testei 5).
		Ficheiro "Método.c" tornado obsoleto.

+++ auxiliar.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
 #include <ctype.h>
 
@@ -21,7 +22,7 @@ int isFloat(char *s){
 		if(isdigit(s[i])) qt++;
 	}
 	
-	if(qt != strlen(s)){
+	if((long unsigned int) qt != strlen(s)){
 		if(strcountc(s, ',') == 0 && strcountc(s, '.') == 0) return 0;
 	}
 	
@@ -29,7 +30,7 @@ int isFloat(char *s){
 		s[strcspn(s, ".")] = ',';
 	}
 	
-	return qt >= strlen(s)-1;
+	return (long unsigned int) qt >= strlen(s)-1;
 }
 
 int isNome(char *s){
@@ -42,7 +43,7 @@ int isNome(char *s){
 		if(isalpha(s[i]) || isspace(s[i])) qt++;
 	}
 	
-	return qt == strlen(s);
+	return (long unsigned int) qt == strlen(s);
 }
 
 int isTel(char *s){
